+++ frontend/src/components/CatDirectoryModal.jsx
@@ -3,7 +3,7 @@ import { useTranslation } from "react-i18next";
 import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
 import { faXmark, faCat } from "@fortawesome/free-solid-svg-icons";
 import { assetUrl, fetchCats } from "../api";
-import { getPosition } from "../lib/geolocate";
+import { getPosition, locateErrorKey } from "../lib/geolocate";
 import { timeAgo } from "../lib/time";
 import Modal from "./Modal";
 import { useToast } from "./Toast";
@@ -41,9 +41,9 @@ export default function CatDirectoryModal({ onClose, onSelect }) {
     }
     getPosition({ highAccuracy: false })
       .then((pos) => setNearCoords({ lat: pos.coords.latitude, lng: pos.coords.longitude }))
-      .catch(() => {
+      .catch((err) => {
         setNearMe(false);
-        toast.error(t("map.locateError"));
+        toast.error(t(locateErrorKey(err)));
       });
   }, [nearMe, toast, t]);